set debug info
