+++ src/extension.ts
@@ -2,13 +2,15 @@
 // Import the module and reference it with the alias vscode in your code below
 
 import * as vscode from 'vscode';
+import {VcpkgLogMgr} from './log';
 import { ConfigurationManager } from './configuration';
 import {SettingsDocument} from './settingsDocument';
 import { VersionManager } from './versionManager';
-import {VcpkgLogMgr} from './log';
 import {CmakeDebugger} from './cmakeDebugger';
 import {VcpkgDebugger} from './vcpkgDebugger';
-import {VcpkgInfoSideBarViewProvider, DepNodeProvider, VcpkgDebuggerSideBarViewProvider} from './sideBar';
+import {VcpkgInfoSideBarViewProvider} from "./sidebar/vcpkgInfoSideBarViewProvider";
+import {VcpkgDebuggerSideBarViewProvider} from './sidebar/vcpkgDebuggerSideBarViewProvider';
+import {DepNodeProvider} from './sidebar/DepNodeProvider';
 
 let logMgr : VcpkgLogMgr;
 let configMgr : ConfigurationManager;
@@ -16,6 +18,8 @@ let verMgr : VersionManager;
 let vcpkgDebugger : VcpkgDebugger;
 let disposables: vscode.Disposable[];
 let cmakeDbg: CmakeDebugger;
+let infoSideBarProvider : VcpkgInfoSideBarViewProvider;
+let debuggerSideBarProvider : VcpkgDebuggerSideBarViewProvider;
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -33,8 +37,8 @@ export function activate(context: vscode.ExtensionContext) {
 	configMgr = new ConfigurationManager(/*context, */verMgr, logMgr, vcpkgDebugger, nodeDependenciesProvider);
 	cmakeDbg = new CmakeDebugger(vcpkgDebugger, logMgr);
 
-	const infoSideBarProvider = new VcpkgInfoSideBarViewProvider(context.extensionUri, context.extensionPath, configMgr);
-	const debuggerSideBarProvider = new VcpkgDebuggerSideBarViewProvider(context.extensionUri, context.extensionPath);
+	infoSideBarProvider = new VcpkgInfoSideBarViewProvider(context.extensionUri, context.extensionPath, configMgr, logMgr);
+	debuggerSideBarProvider = new VcpkgDebuggerSideBarViewProvider(context.extensionUri, context.extensionPath, vcpkgDebugger, logMgr);
 	
 	configMgr.logInfo('Trying to active vcpkg plugin...');
